% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_all_funs.R
\name{aj_py}
\alias{aj_py}
\title{create interactive python repl on load}
\usage{
aj_py()
}
\arguments{
\item{aj_py}{launch interactive python repl}
}
\description{
create interactive python repl on load
}
