% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_all_funs.R
\name{ajh}
\alias{ajh}
\title{html package help function}
\usage{
ajh(...)
}
\arguments{
\item{...}{usually \code{library = "package-name"}}
}
\description{
A collection of useful functions
}
\details{
This is an initial package development with my functions. The plan will be
to build on these and separate functions out into seperate documents.
Additionally, automated testing may be added at a later date.
}
\author{
Atanas Janackovski
}
