% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_all_funs.R
\name{coalesce_by_column}
\alias{coalesce_by_column}
\title{function to collapse duplicates entries}
\usage{
coalesce_by_column(df)
}
\arguments{
\item{df}{this function should be run on a dataframe or matrix}
}
\description{
function to collapse duplicates entries
}
\examples{
## collapse duplicates and recode any with problems
\dontrun{
sp_change <-
    sp_change \%>\%
    group_by(uci) \%>\%
    summarise_all(coalesce_by_column) \%>\%
    ungroup()
}
}
