% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_all_funs.R
\name{corstars}
\alias{corstars}
\title{Create correlation tables for printing}
\usage{
corstars(
  data,
  method = c("pearson", "spearman", "kendall"),
  removeTriangle = c("upper", "lower"),
  result = c("none", "html")
)
}
\arguments{
\item{data}{your data}

\item{method}{correlation methods taken from "psych::corr.test()", defaults to "pearson". # nolint}

\item{removeTriangle}{whic triangle do you want emptied? Defaults to "upper". # nolint}

\item{result}{how do you want your starts wrapped? Defaults to "none".}
}
\description{
This is a function to create correlation tables. The expecation
is that this will be passed meting linke \code{kable}. Code adpated from
\href{http://sthda.com/english/wiki/elegant-correlation-table-using-xtable-r-package}{here}. # nolint
}
