% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_all_funs.R
\name{descBy_df}
\alias{descBy_df}
\title{create df of descriptives by group}
\usage{
descBy_df(df, grp, ...)
}
\arguments{
\item{df}{your data}

\item{grp}{var to group by}

\item{...}{accepts arguments used as per \code{pscyh::describeBy}}
}
\description{
create df of descriptives by group
}
