% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_all_funs.R
\name{outlier_boxplots}
\alias{outlier_boxplots}
\title{multiple boxplots with outliers}
\usage{
outlier_boxplots(data, melt_var, incl = incl, excl = excl)
}
\arguments{
\item{data}{your data}

\item{melt_var}{var to melt by}

\item{incl}{requires regex string of var names to include}

\item{excl}{requires regex string of var names to exclude}
}
\description{
multiple boxplots with outliers
}
