% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_all_funs.R
\name{pro_mean_miss}
\alias{pro_mean_miss}
\title{create prorated means and percent missing of selected variables}
\usage{
pro_mean_miss(data, new_name, ...)
}
\arguments{
\item{data}{your data}

\item{new_name}{the prefix of the new variables as a string}

\item{...}{variables to select as per dplyr::select}
}
\description{
create prorated means and percent missing of selected variables
}
