% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_all_funs.R
\name{py}
\alias{py}
\title{create interactive python repl on load}
\usage{
py(module = NULL, quiet = getOption("reticulate.repl.quiet", default = FALSE))
}
\description{
create interactive python repl on load
}
