% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_all_funs.R
\name{qq_plots}
\alias{qq_plots}
\title{create numerous qq_plots of selected variables}
\usage{
qq_plots(df)
}
\arguments{
\item{df}{your data}
}
\description{
create numerous qq_plots of selected variables
}
\details{
use with dplyr::select or purr::keep

https://drsimonj.svbtle.com/quick-plot-of-all-variables
}
