% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_all_funs.R
\name{row_means_miss_perc}
\alias{row_means_miss_perc}
\title{create row means and percent missing of selected variables}
\usage{
row_means_miss_perc(data, new_name, ...)
}
\arguments{
\item{data}{your data}

\item{new_name}{the prefix of the new variables as a string}

\item{...}{variables to select as per dplyr::select}
}
\description{
create row means and percent missing of selected variables
}
