% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_all_funs.R
\name{workspace_size}
\alias{workspace_size}
\title{function returns total size of all objects in current env}
\usage{
workspace_size()
}
\arguments{
\item{workspace_size}{go get it}
}
\description{
function returns total size of all objects in current env
}
\examples{
## just run the function and get the size
\dontrun{
workspace_size()
}
}
